working on file overhaul so its cleaner
