+++ src/game/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(SOURCES
+   src/game/Player.h
+   src/game/Player.c
+
+    PARENT_SCOPE)
