+++ src/math/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(SOURCES
+   src/math/Vector.h
+   src/math/Vector.c
+
+    PARENT_SCOPE)
+
