+++ src/window/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(SOURCES
+   src/window/Window.h
+   src/window/Window.c
+
+    PARENT_SCOPE)
+
